crop_agent/requirements.txt
python-dotenv
requests
gTTS
playsound